repaired url's, added home page, made url's private from other logged in users, not fully checked for bugs though

+++ home/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render
+from django.http import Http404
 from django.template import RequestContext
 from django.contrib.auth.models import User
 from django.http import HttpResponseRedirect
@@ -7,8 +8,20 @@ from django.shortcuts import reverse,redirect, get_object_or_404
 from home.models import Document,Group
 from home.forms import DocumentForm, GroupRegistrationForm
 
+def home(request):
+    user=request.user
+    context = {'user':user}
+    return render(
+        request, 'home/templates/home.html', context
+    )
 def group_detail_files(request,id=None):
     group=get_object_or_404(Group, id=id)
+    thief=True
+    for u in group.members.all():
+        if u==request.user:
+            thief=False
+    if thief:
+        raise Http404
     document=Document.objects.all()
     context = {'group':group,'document':document}
     return render(
@@ -17,6 +30,13 @@ def group_detail_files(request,id=None):
 def group_detail_files_upload(request,id=None):
     # Handle file upload
     group = get_object_or_404(Group, id=id)
+    thief = True
+    for u in group.members.all(x):
+        if u == request.user:
+            thief = False
+    if thief:
+        raise Http404
+
     if request.method == 'POST':
         form = DocumentForm(request.POST, request.FILES)
         print(str(form.errors))
@@ -36,11 +56,19 @@ def group_detail_files_upload(request,id=None):
     return render( request,'home/templates/home_groups_specified_files_upload.html',context )
 
 def group_detail_file_delete(request,id=None,idf=None):
-    instanceF=get_object_or_404(Document,id=idf)
-    instanceG=get_object_or_404(Group,id=id)
+    instanceF = get_object_or_404(Document, id=idf)
+    instanceG = get_object_or_404(Group, id=id)
+    thief = True
+    for u in instanceG.members.all():
+        if u == request.user:
+            thief = False
+    if thief:
+        raise Http404
+
     instanceF.delete()
     return redirect(instanceG.get_absolute_url())
 def group_list(request):
+
     groups = Group.objects.all()
     user=request.user
     context = {'groups': groups,'user':user}
@@ -49,6 +77,13 @@ def group_list(request):
     )
 def group_edit(request,id=None):
     instance=get_object_or_404(Group,id=id)
+
+    thief = True
+    for u in instance.members.all():
+        if u == request.user:
+            thief = False
+    if thief:
+        raise Http404
     form= GroupRegistrationForm(request.POST or None,instance=instance)
     if form.is_valid():
         instance=form.save()
@@ -61,6 +96,12 @@ def group_edit(request,id=None):
 
 def group_detail(request,id=None):
     instance=get_object_or_404(Group,id=id)
+    thief = True
+    for u in instance.members.all():
+        if u == request.user:
+            thief = False
+    if thief:
+        raise Http404
     context = {'id':id,'name': instance.name, 'members': instance.members}
     return render(
         request, 'home/templates/home_groups_specified.html', context)
